have step size change when sending blocks
To get motors to run smoothly, change the delta step size based on the
feed rate instead of holding it fixed at .1.  Do not see motors changing
speed any longer.  When held at .1, it seemed to be too small for faster
feeds.  Too many blocks were needed per sec, and seemed to be a
bottleneck.