PID to control blob size in front of a die
Added another PID loop to control the blob size in front of a die.  The
PID loop controls the extruder motor.  This requires an additional
sensor to measure the blob.  Sensor input is expected on Analog input 3
of the Mega, wich is a pin on AUX 1.  LCD menus added.  Also, the
settings are saved in EEPROM.  This change will reset EEPROM settings to
default.